SMPP: Use 32-bit integer for PDU length to prevent truncation
According to SMPP v3.4 and v5.0 specifications (section 4.1.1), the command_length field is a 4-octet integer value. The dissector read the length using tvb_get_ntohl() but assigned it to a uint16_t variable, causing truncation for PDU lengths exceeding 65535 bytes.

This caused the dissector to either prematurely abort dissection when the lower 16 bits wrapped to zero or desynchronize from PDU boundaries and parse payload contents as headers.

Change pdu_len to uint32_t to match the 4-byte wire representation.

+++ epan/dissectors/packet-smpp.c
@@ -2833,7 +2833,7 @@ dissect_smpp(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree, void* data)
     else {                            /* no? probably X.25            */
         uint32_t offset = 0;
         while (tvb_reported_length_remaining(tvb, offset) > 0) {
-            uint16_t pdu_len = tvb_get_ntohl(tvb, offset);
+            uint32_t pdu_len = tvb_get_ntohl(tvb, offset);
             int pdu_real_len = tvb_captured_length_remaining(tvb, offset);
             tvbuff_t *pdu_tvb;
